+++ src/zod-plugin.ts
@@ -10,7 +10,7 @@ import { clone, fromPairs, map, pipe, toPairs, pair } from "ramda";
 import { z } from "zod";
 import { FlatObject } from "./common-helpers";
 import { cloneSchema, Metadata, metaSymbol } from "./metadata";
-import { Remap } from "./mapping-helpers";
+import { Intact, Remap } from "./mapping-helpers";
 
 declare module "zod" {
   interface ZodTypeDef {
@@ -31,11 +31,11 @@ declare module "zod" {
     Output = z.objectOutputType<T, Catchall, UnknownKeys>,
     Input = z.objectInputType<T, Catchall, UnknownKeys>,
   > {
-    remap<V extends string, U extends { [P in keyof T]: V }>(
+    remap<V extends string, U extends { [P in keyof T]?: V }>(
       mapping: U,
     ): z.ZodPipeline<
       z.ZodEffects<this, FlatObject>, // internal type simplified
-      z.ZodObject<Remap<T, U, V>>
+      z.ZodObject<Remap<T, U, V> & Intact<T, U>>
     >;
   }
 }
@@ -75,17 +75,19 @@ const objectMapper = function (
   return this.transform(
     pipe(
       toPairs,
-      map(([key, value]) => pair(mapping[key], value)),
+      map(([key, value]) => pair(mapping[key] || key, value)),
       fromPairs,
     ),
   ).pipe(
-    z.object(
-      pipe(
-        toPairs,
-        map(([key, schema]) => pair(mapping[String(key)], schema)),
-        fromPairs,
-      )(clone(this.shape)), // immutable, no references to the original schemas
-    ),
+    z
+      .object(
+        pipe(
+          toPairs,
+          map(([key, schema]) => pair(mapping[String(key)] || key, schema)),
+          fromPairs,
+        )(clone(this.shape)), // immutable, no references to the original schemas
+      )
+      [this._def.unknownKeys](), // proxies unknown keys when set to "passthrough"
   );
 };