+++ tests/unit/zod-plugin.spec.ts
@@ -81,5 +81,32 @@ describe("Zod Runtime Plugin", () => {
         userId: "test",
       });
     });
+
+    test.each([{ user_id: "userId" }, { user_id: "userId", name: undefined }])(
+      "should support partial mapping %#",
+      (mapping) => {
+        const schema = z.object({ user_id: z.string(), name: z.string() });
+        const mappedSchema = schema.remap(mapping);
+        expect(mappedSchema._def.out.shape).toEqual({
+          userId: schema.shape.user_id,
+          name: schema.shape.name,
+        });
+        expect(mappedSchema.parse({ user_id: "test", name: "some" })).toEqual({
+          userId: "test",
+          name: "some",
+        });
+      },
+    );
+
+    test("should support passthrough object schemas", () => {
+      const schema = z.object({ user_id: z.string() }).passthrough();
+      const mappedSchema = schema.remap({ user_id: "userId" });
+      expect(
+        mappedSchema.parse({ user_id: "test", extra: "excessive" }),
+      ).toEqual({
+        userId: "test",
+        extra: "excessive",
+      });
+    });
   });
 });